+++ CHANGELOG.md
@@ -13,6 +13,7 @@ CHANGELOG for FlatCAM Evo beta
 - PDF Tool: refactoring the tool into its own folder
 - ParsePDF.py: refactoring the parser to use numpy for the bezier_to_circle method
 - ParseGerber: optimization of the geometry creation process
+- ParseGerber: optimization of the geometry creation process - part 2
 
 8.10.2025